Ejemplo de una clase generica, con esto podemos hacer uso de mas genericos

+++ src/main/kotlin/kotlin_avanzado/genericos/Dimension.kt
@@ -1,3 +1,3 @@
 package kotlin_avanzado.genericos
 
-data class Dimension(val x: Float, val y: Float, val z: Float)
+data class Dimension(val x: Double, val y: Double, val z: Double)
